More changes for edit future years.
