+++ forecast/views/view_forecast/export_forecast_data.py
@@ -2,6 +2,8 @@ from django.contrib.auth.decorators import user_passes_test
 from django.shortcuts import redirect
 from django.urls import reverse
 
+from core.utils.generic_helpers import get_current_financial_year
+
 from forecast.models import (
     ForecastingDataView,
 )
@@ -216,13 +218,16 @@ def export_forecast_data_project_detail_dit(request, project_code_id, period):
     return export_forecast_data_generic(period, filter, title)
 
 
-def export_edit_forecast_data(request, cost_centre):
+def export_edit_forecast_data(request, cost_centre, financial_year = 0):
     fields = ForecastQueryFields()
     if can_edit_cost_centre(request.user, cost_centre):
+        if financial_year == 0:
+            financial_year = get_current_financial_year()
         filter = {fields.cost_centre_code_field: cost_centre}
         q = ForecastingDataView.view_data.raw_data_annotated(
             {**fields.EDIT_KEYS_DOWNLOAD, **fields.EDIT_FORECAST_DOWNLOAD_COLUMNS},
             filter,
+            financial_year,
             order_list=fields.EDIT_FORECAST_DOWNLOAD_ORDER,
         )
         title = f"Edit forecast {cost_centre}"
